level 3(ish)
I have it so you can search for any bookings that contain a certain date, be it the check in date or the check out date. I'm unclear on the instruction as it seems the user would only search one date - I'm unclear on how I should span dates when I only have one to work with.